proyecto final

+++ Proyecto-UserManager/admin/delete.php
@@ -1,16 +1,20 @@
 <?php
-
 session_start();
-require_once '../db.php';
-
 
 if (!isset($_SESSION['user_id']) || $_SESSION['user_rol'] !== 'admin') {
     header("Location: ../login.php");
     exit();
 }
 
+require_once '../db.php';
+
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
+
+    if ($id == $_SESSION['user_id']) {
+        header("Location: index.php?error=no_te_borres");
+        exit();
+    }
 
     try {
         $stmt = $pdo->prepare("DELETE FROM usuarios WHERE id = ?");
